pyproject.toml
[project]
name = "wrdz"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Pedro Werneck", email = "pjwerneck@gmail.com" }
]
requires-python = ">=3.8"
dependencies = [
    "msgpack>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "hypothesis>=6.113.0",
    "pytest>=8.3.4",
    "pytest-benchmark>=4.0.0",
    "pytest-coverage>=0.0",
    "rich>=13.9.4",
    "ruff>=0.9.4",
    "smaz-py3>=1.1.3",
]
[tool.ruff]
exclude = [".git", ".venv"]
line-length = 99
target-version = "py312"
extend-exclude = ["neuropad/core/**"]

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint.isort]
force-single-line = true
